Fix docs build and add code examples
